+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo.hs
@@ -30,13 +30,17 @@ import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.Era
 import Cardano.Ledger.Alonzo.Forecast ()
 import Cardano.Ledger.Alonzo.PParams ()
-import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusContext, LedgerTxInfo (..))
+import Cardano.Ledger.Alonzo.Plutus.Context (
+  EraPlutusContext,
+  LedgerTxInfo (..),
+  SupportedPlutusRunnable (..),
+ )
 import Cardano.Ledger.Alonzo.Plutus.Evaluate (
   scriptsWithContextFromLedgerTxInfo,
  )
 import Cardano.Ledger.Alonzo.Plutus.TxInfo ()
 import Cardano.Ledger.Alonzo.Rules ()
-import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..), plutusScriptLanguage)
+import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..))
 import Cardano.Ledger.Alonzo.State ()
 import Cardano.Ledger.Alonzo.Transition ()
 import Cardano.Ledger.Alonzo.Translation ()
@@ -52,7 +56,7 @@ import Cardano.Ledger.Alonzo.UTxO (
 import Cardano.Ledger.Binary (DecCBOR, EncCBOR)
 import Cardano.Ledger.Block (EraBlockHeader)
 import Cardano.Ledger.Mary.Value (MaryValue)
-import Cardano.Ledger.Plutus.Data ()
+import Cardano.Ledger.Plutus (plutusLanguage)
 import Cardano.Ledger.Rules.ValidationMode (lblStatic)
 import Cardano.Ledger.Shelley.API
 import Cardano.Ledger.Shelley.Rules (ledgerPpL, ledgerSlotNoL)
@@ -126,12 +130,13 @@ mkAlonzoStAnnTx ::
   AlonzoStAnnTx TopTx era
 mkAlonzoStAnnTx ei sysStart pp utxo tx =
   let
+    protVer = pp ^. ppProtocolVersionL
     scriptsNeeded = getScriptsNeeded utxo (tx ^. bodyTxL)
     scriptsProvided = getScriptsProvided utxo tx
-    plutusScriptsUsed = resolveNeededPlutusScriptsWithPurpose scriptsProvided scriptsNeeded
+    plutusScriptsUsed = resolveNeededPlutusScriptsWithPurpose protVer scriptsProvided scriptsNeeded
     ledgerTxInfo =
       LedgerTxInfo
-        { ltiProtVer = pp ^. ppProtocolVersionL
+        { ltiProtVer = protVer
         , ltiEpochInfo = ei
         , ltiSystemStart = sysStart
         , ltiUTxO = utxo
@@ -144,7 +149,7 @@ mkAlonzoStAnnTx ei sysStart pp utxo tx =
       , asatScriptsNeeded = scriptsNeeded
       , asatScriptsProvided = scriptsProvided
       , asatPlutusLanguagesUsed =
-          Set.fromList [plutusScriptLanguage s | (_, _, s) <- plutusScriptsUsed]
+          Set.fromList [plutusLanguage spr | (_, SupportedPlutusRunnable spr) <- plutusScriptsUsed]
       , asatPlutusScriptsWithContext =
           scriptsWithContextFromLedgerTxInfo ledgerTxInfo (pp ^. ppCostModelsL) plutusScriptsUsed
       }