+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Plutus/Context.hs
@@ -29,6 +29,7 @@ module Cardano.Ledger.Alonzo.Plutus.Context (
   toPlutusTxInfoForPurpose,
   EraPlutusContext (..),
   lookupTxInfoResultImpossible,
+  SupportedPlutusRunnable (..),
   SupportedLanguage (..),
   mkSupportedLanguageM,
   supportedLanguages,
@@ -52,21 +53,19 @@ import Cardano.Ledger.Alonzo.Scripts (
   PlutusPurpose,
   PlutusScript (..),
  )
-import Cardano.Ledger.BaseTypes (ProtVer (..), kindObjectValue)
+import Cardano.Ledger.BaseTypes (ProtVer (..), Version, kindObjectValue)
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Core
 import Cardano.Ledger.Plutus (
-  CostModel,
   Data,
-  ExUnits,
   Language (..),
   Plutus (..),
   PlutusArgs,
   PlutusBinary,
   PlutusLanguage,
+  PlutusRunnable,
   PlutusScriptContext,
-  PlutusWithContext (..),
   SLanguage (..),
   asSLanguage,
   plutusLanguage,
@@ -204,6 +203,8 @@ class
 
   mkSupportedLanguage :: Language -> Maybe (SupportedLanguage era)
 
+  mkSupportedPlutusRunnable :: Version -> PlutusScript era -> SupportedPlutusRunnable era
+
   -- | Construct `PlutusTxInfo` for all supported languages in this era.
   mkTxInfoResult :: LedgerTxInfo era -> TxInfoResult era
 
@@ -217,16 +218,6 @@ class
     SLanguage l ->
     TxInfoResult era ->
     PlutusTxInfoResult l era
-
-  mkPlutusWithContext ::
-    PlutusScript era ->
-    ScriptHash ->
-    PlutusPurpose AsIxItem era ->
-    LedgerTxInfo era ->
-    TxInfoResult era ->
-    (Data era, ExUnits) ->
-    CostModel ->
-    Either (ContextError era) PlutusWithContext
 
 -- | Helper function to use when implementing `lookupTxInfoResult` for plutus languages that are not
 -- supported by the era.
@@ -296,6 +287,11 @@ mkSupportedBinaryPlutusScript supportedLanguage plutus =
   case supportedLanguage of
     SupportedLanguage sLang ->
       mkSupportedPlutusScript (asSLanguage sLang (Plutus plutus))
+
+data SupportedPlutusRunnable era where
+  SupportedPlutusRunnable ::
+    EraPlutusTxInfo l era =>
+    !(PlutusRunnable l) -> SupportedPlutusRunnable era
 
 data SupportedLanguage era where
   SupportedLanguage :: EraPlutusTxInfo l era => SLanguage l -> SupportedLanguage era